+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2020-08-16
+
+### Added
+- Repository module for REST calls
+- ViewModels for persisting UI state
+
+### Fixed
+- Scrolling the text of articles
+- NPE when an article has an empty body
+
+
 ## [0.6.0] - 2020-08-12
 
 ### Added